data/UNCTAD/requirements_load.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
numpy>=1.21.0